# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deck OS - JARVIS Cyberdeck Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai-router/infer:
    post:
      operationId: routeInference
      tags: [ai-router]
      summary: Route an inference request to the best available model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InferenceRequest"
      responses:
        "200":
          description: Inference response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InferenceResponse"
  /ai-router/mode:
    get:
      operationId: getIntelligenceMode
      tags: [ai-router]
      summary: Get current intelligence mode
      responses:
        "200":
          description: Current intelligence mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceMode"
    put:
      operationId: setIntelligenceMode
      tags: [ai-router]
      summary: Set intelligence mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetIntelligenceModeRequest"
      responses:
        "200":
          description: Updated intelligence mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceMode"
...
  /memory/short-term:
    get:
      operationId: getShortTermMemory
      tags: [memory]
      summary: Get current session memory
      responses:
        "200":
          description: Short-term session memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryStore"
    post:
      operationId: storeShortTermMemory
      tags: [memory]
      summary: Store a new short-term memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEntryCreate"
      responses:
        "201":
          description: Memory stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /memory/long-term:
    get:
      operationId: getLongTermMemory
      tags: [memory]
      summary: Search long-term persistent memory
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Long-term memory results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryStore"
    post:
      operationId: storeLongTermMemory
      tags: [memory]
      summary: Store a new long-term memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEntryCreate"
      responses:
        "201":
          description: Memory stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /memory/long-term/{id}:
    delete:
      operationId: deleteMemoryEntry
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Memory deleted
...
  /devices/{deviceId}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device details and sensor readings
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{deviceId}/control:
    post:
      operationId: controlDevice
      tags: [devices]
      summary: Send a control command to a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceControlRequest"
      responses:
        "200":
          description: Control result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceControlResult"
...
components:
  schemas:
...
    InferenceRequest:
      type: object
      properties:
        prompt:
          type: string
        mode:
          type: string
          enum: [fast, deep, none]
        pluginId:
          type: string
          nullable: true
        context:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required:
              - role
              - content
        useCache:
          type: boolean
          default: true
      required:
        - prompt
        - mode
    InferenceResponse:
      type: object
      properties:
        response:
          type: string
        modelUsed:
          type: string
        mode:
          type: string
        fromCache:
          type: boolean
        latencyMs:
          type: integer
        tokens:
          type: integer
          nullable: true
      required:
        - response
        - modelUsed
        - mode
        - fromCache
        - latencyMs
    IntelligenceMode:
      type: object
      properties:
        mode:
          type: string
          enum: [DIRECT_EXECUTION, LIGHT_REASONING, DEEP_REASONING, HYBRID_MODE]
        description:
          type: string
        preferLocal:
          type: boolean
      required:
        - mode
        - description
        - preferLocal
    SetIntelligenceModeRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [DIRECT_EXECUTION, LIGHT_REASONING, DEEP_REASONING, HYBRID_MODE]
      required:
        - mode
...
    MemoryStore:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/MemoryEntry"
        total:
          type: integer
      required:
        - entries
        - total
    MemoryEntry:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [short_term, long_term]
        content:
          type: string
        keywords:
          type: array
          items:
            type: string
        source:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - type
        - content
        - keywords
        - source
        - createdAt
    MemoryEntryCreate:
      type: object
      properties:
        content:
          type: string
        keywords:
          type: array
          items:
            type: string
        source:
          type: string
        ttlSeconds:
          type: integer
          nullable: true
      required:
        - content
        - keywords
        - source
...
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [sensor, actuator, display, network, simulated]
        status:
          type: string
          enum: [online, offline, error, standby]
        protocol:
          type: string
          enum: [mqtt, websocket, http, simulated]
        readings:
          type: array
          items:
            $ref: "#/components/schemas/DeviceReading"
        capabilities:
          type: array
          items:
            type: string
        lastSeen:
          type: string
          format: date-time
          nullable: true
        location:
          type: string
          nullable: true
      required:
        - id
        - name
        - type
        - status
        - protocol
        - readings
        - capabilities
    DeviceReading:
      type: object
      properties:
        sensor:
          type: string
        value:
          oneOf:
            - type: number
            - type: string
            - type: boolean
        unit:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - sensor
        - value
        - timestamp
    DeviceControlRequest:
      type: object
      properties:
        action:
          type: string
        parameters:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - action
    DeviceControlResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        newState:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - success
        - message
...